src/requirements.txt
aio-pika==9.4.1
aiohttp==3.9.3
aiohttp-retry==2.8.3
aioredis==2.0.1
aiormq==6.8.0
aiosignal==1.3.1
amqp==5.2.0
anyio==4.2.0
apiview==1.3.25
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.0
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
beautifulsoup4==4.12.2
billiard==4.2.0
black==24.3.0
build==0.10.0
bunch==1.0.1
CacheControl==0.13.1
cachetools==5.3.3
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
cleo==2.0.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
configparser==5.3.0
constantly==23.10.4
coreapi==2.3.3
coreschema==0.0.4
coverage==7.4.0
crashtest==0.4.1
cron-descriptor==1.4.3
cryptography==42.0.5
cssselect==1.2.0
cssutils==2.7.1
daphne==4.1.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
diff-match-patch==20200713
distlib==0.3.7
Django==5.0.3
django-braces==1.15.0
django-celery-beat==2.6.0
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-dotenv==1.4.2
django-extensions==3.2.3
django-filter==23.2
django-import-export==3.2.0
django-jazzmin==2.6.0
django-js-asset==2.2.0
django-mptt==0.16.0
django-oauth-toolkit==2.3.0
django-redis==5.4.0
django-resized==1.0.2
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-rest-framework-social-oauth2==1.1.0
django-rest-multiple-models==2.1.3
django-rest-swagger==2.2.0
django-simple-history==3.5.0
django-timezone-field==6.1.0
django-use-email-as-username==1.4.0
Django-Verify-Email==2.0.3
djangochannelsrestframework==1.2.0
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
drf-social-oauth2==2.1.3
drf-writable-nested==0.7.0
drf-yasg==1.21.7
dulwich==0.21.6
dynaconf==3.1.12
ecdsa==0.18.0
et-xmlfile==1.1.0
exceptiongroup==1.2.0
facebook-sdk==3.1.0
fastjsonschema==2.19.0
filelock==3.12.4
firebase-admin==6.4.0
frozenlist==1.4.1
google==3.0.0
google-api-core==2.15.0
google-api-python-client==2.114.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.15.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpcio==1.62.0
grpcio-status==1.62.0
gunicorn==21.2.0
h11==0.14.0
httplib2==0.22.0
httptools==0.6.1
humanize==4.9.0
hvac==1.1.0
hyperlink==21.0.0
idna==3.6
importlib-metadata==6.8.0
incremental==22.10.0
inflection==0.5.1
iniconfig==2.0.0
install==1.3.5
installer==0.7.0
itypes==1.2.0
jaraco.classes==3.3.0
jeepney==0.8.0
Jinja2==3.1.3
json2xml==4.1.0
jsonschema==4.17.3
jwcrypto==1.5.0
keyring==24.2.0
kombu==5.3.5
loguru==0.7.0
lxml==4.9.2
MarkupPy==1.14
MarkupSafe==2.1.5
mock==5.0.2
more-itertools==10.1.0
msgpack==1.0.8
multidict==6.0.5
mypy-extensions==1.0.0
oauthlib==3.2.2
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.1.2
packaging==24.0
pamqp==3.3.0
pathspec==0.12.1
peewee==3.17.0
pexpect==4.8.0
pika==1.3.2
pillow==10.2.0
pkginfo==1.9.6
platformdirs==3.9.1
pluggy==1.3.0
poetry==1.6.1
poetry-core==1.7.0
poetry-plugin-export==1.5.0
premailer==3.10.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.2
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyhcl==0.4.4
PyJWT==2.8.0
pyOpenSSL==24.1.0
pyotp==2.9.0
pyparsing==3.1.1
pyproject_hooks==1.0.0
pyrsistent==0.19.3
pytest==7.4.4
pytest-cov==2.12.1
python-crontab==3.0.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-jose==3.3.0
python-json-logger==2.0.7
python3-openid==3.2.0
pytz==2024.1
pywin32-ctypes==0.2.2
PyYAML==6.0.1
rand==0.3.0
rapidfuzz==2.15.2
readonly==0.1.2
redis==5.0.3
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rest-framework-simplejwt==0.0.2
rsa==4.9
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
SecretStorage==3.3.3
service-identity==24.1.0
shellingham==1.5.4
shortuuid==1.0.11
simple-configparser==0.1.4.0
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
soupsieve==2.5
sqlparse==0.4.4
tablib==3.4.0
templated-mail==0.4.0.2
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.1
trove-classifiers==2023.10.18
twilio==9.0.2
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.27.1
validators==0.20.0
vine==5.1.0
virtualenv==20.24.6
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
win32-setctime==1.1.0
wrapt==1.15.0
xlrd==2.0.1
xlwt==1.3.0
xmltodict==0.13.0
yagmail==0.15.293
yarl==1.9.4
zipp==3.17.0
zope.interface==6.2
